+++ jest.config.js
@@ -1,0 +1,21 @@
+export default {
+  preset: 'ts-jest/presets/default-esm',
+  extensionsToTreatAsEsm: ['.ts'],
+  testEnvironment: 'node',
+  roots: ['<rootDir>/tests'],
+  testMatch: ['**/*.jest.test.ts'],
+  transform: {
+    '^.+\.ts$': ['ts-jest', {
+      useESM: true
+    }],
+  },
+  collectCoverageFrom: [
+    'src/**/*.ts',
+    '!src/**/*.d.ts',
+  ],
+  coverageDirectory: 'coverage-jest',
+  coverageReporters: ['text', 'lcov', 'html'],
+  setupFilesAfterEnv: [],
+  moduleFileExtensions: ['ts', 'js', 'json'],
+  verbose: true,
+};